orientation-event/DeviceOrientation/t012.html

<!DOCTYPE html>
<html>

  <head>
    <title>Implementations must set the values of the unknown angles to null:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" /> 
    <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
    <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  		<h1>Implementations must set the values of the unknown angles to null:</h1>
    <div id="log">
    </div>
	
	<script>
  		window.addEventListener('deviceorientation', function(eventData) {

			// Gamma : angle par rapport a x
			gamma = eventData.gamma;
			// Beta : angle par rapport a y
			beta = eventData.beta;
			// Alpha : orientation (N-S-E-O)
			alpha = eventData.alpha;

		}, false);
	
	alert("Launch Test.");
	
	test(function() {assert_true(gamma == null)}, "check if gamma is set to null");
	test(function() {assert_true(beta == null)}, "check if beta is set to null");
	test(function() {assert_true(alpha == null)}, "check if alpha is set to null");
		
	</script>
	
  </body>
  
</html>
